services/api/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.14.0
neo4j==5.27.0

# Redis
redis==5.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Encryption (credential storage at rest)
cryptography==43.0.0

# HTTP
httpx[sse]==0.27.2
requests==2.32.3

# LLM / AI
anthropic==0.38.0
openai==1.52.0

# Network
netmiko==4.4.0
napalm==5.1.0
easysnmp==0.2.6

# Vector DB
pgvector==0.3.0

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.3.0

# Utilities
python-multipart==0.0.12
python-dotenv==1.0.1
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Type checking
mypy==1.14.0
slowapi==0.1.9
